# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Hedera Starter Project",
    "description": "Hashgraph Interaction API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/hedera": {
      "get": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Get Hedera Account in system",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/{initialBalance}": {
      "get": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Get Hedera Account in system",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Create an account on hedera with certain balance (transferred from default account).",
        "parameters": [
          {
            "name": "initialBalance",
            "in": "path",
            "description": "add this much tinyBars to the user.",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract": {
      "post": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Deploy a contract on hedera.",
        "parameters": [
          {
            "name": "Request Body",
            "in": "body",
            "description": "Request Body",
            "schema": {
              "type": "object",
              "properties": {
                "bytecodeFileId": {
                  "type": "string"
                },
                "tokenAddressSol": {
                  "type": "string"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/addBytecode": {
      "post": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Add the Bytecode File to Hedera (and Create a Token!).",
        "parameters": [
          {
            "name": "bytecode",
            "in": "body",
            "description": "bytecode",
            "schema": {
              "type": "object",
              "properties": {
                "bytecode": {
                  "type": "string"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/{contractId}": {
      "get": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Get info a contract on hedera.",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "Contract id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Delete a contract on hedera.",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "Contract id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/{contractId}/bytecode" : {
      "get": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Get bytecode a contract on hedera.",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "Contract id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/{contractId}/state-size" : {
      "get": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Get state size a contract on hedera.",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "Contract id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/execute-transaction": {
      "post": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Execute a transaction on a contract that takes in a string argument.",
        "parameters": [
          {
            "name": "requestDTO",
            "in": "body",
            "description": "DTO to execute a transaction on a contract",
            "schema": {
              "type": "object",
              "properties": {
                "contractId": {
                  "type": "string",
                  "default": "0.0.3388093"
                },
                "functionName": {
                  "default": "mintFungibleToken",
                  "type": "string"
                },
                "argument": {
                  "type": "number",
                  "default": "150"
                },
                "gasValue": {
                  "type": "number",
                  "default": 3000000
                },
                "tokenId": {
                  "type": "string",
                  "default": "0.0.3387432"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contract/call-method": {
      "post": {
        "tags": [
          "Hedera Contract"
        ],
        "summary": "Call a method on a contract that takes in a string argument and returns string.",
        "parameters": [
          {
            "name": "requestDTO",
            "in": "body",
            "description": "DTO to execute a transaction on a contract",
            "schema": {
              "type": "object",
              "properties": {
                "contractId": {
                  "type": "string"
                },
                "functionName": {
                  "type": "string"
                },
                "argument": {
                  "type": "string"
                },
                "gasValue": {
                  "type": "number"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/info/{accountId}": {
      "get": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Get account info",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account": {
      "delete": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Delete Hedera account",
        "parameters": [
          {
            "name": "requestDTO",
            "in": "body",
            "description": "DTO to execute a deleting account transaction",
            "schema": {
              "type": "object",
              "properties": {
                "accountId": {
                  "type": "string"
                },
                "accountPrivateKey": {
                  "type": "string"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/balance/{accountId}": {
      "get": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Get balance of hedera account",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/transaction/cryptoTransfer": {
      "post": {
        "tags": [
          "Hedera Account"
        ],
        "summary": "Transfer Hbars between 2 accounts.",
        "parameters": [
          {
            "name": "requestDTO",
            "in": "body",
            "description": "DTO to execute a transaction on an account",
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "number",
                  "required": [true]
                },
                "memo": {
                  "type": "string"
                },
                "senderId": {
                  "type": "string"
                },
                "senderPrivateKey": {
                  "type": "string"
                },
                "receiverId": {
                  "type": "string",
                  "required": [true]
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/topic": {
      "post": {
        "tags": [
          "HCS"
        ],
        "summary": "Create a HCS Topic",
        "responses": {
          "200": {
            "description": "Topic ID",
            "examples": {
              "Topic ID": "00000.111"
            }
          }
        }
      }
    },
    "/submitMessage": {
      "post": {
        "tags": [
          "HCS"
        ],
        "summary": "Submit a message to a topic",
        "parameters": [
          {
            "in": "query",
            "name": "topicId",
            "type": "string",
            "default": "0.0.3120640",
            "description": "0.0.3120640"
          },
          {
            "in": "query",
            "name": "message",
            "type": "string",
            "default": "Hello Duy!",
            "description": "test"
          }
        ],
        "responses": {
          "200": {
            "description": "success or failure"
          }
        }
      }
    },
    "/subscribe": {
      "post": {
        "tags": [
          "HCS"
        ],
        "summary": "Subscribe Mirror Node to a topicId. Updates to topics will be printed to the console.",
        "parameters": [
          {
            "in": "query",
            "name": "topicId",
            "type": "string",
            "default": "0.0.3120640",
            "description": "0.0.3120640"
          }
        ],
        "responses": {
          "200": {
            "description": "success or failure"
          }
        }
      }
    },
    "/topic/{topicId}": {
      "get": {
        "tags": [
          "HCS"
        ],
        "summary": "Get info on a topic",
        "parameters": [
          {
            "name": "topicId",
            "in": "path",
            "description": "topicId",
            "type": "string",
            "default": "0.0.2117",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "TopicInfo",
            "schema": {
              "$ref": "#/definitions/TopicInfo"
            }
          }
        }
      }
    },
    "/file/create": {
      "post": {
        "tags": [
          "File controller"
        ],
        "summary": "Store the Smart Contract Bytecode on Hedera.",
        "parameters": [
          {
            "in": "query",
            "required": true,
            "name": "text",
            "type": "string",
            "default": "Testing",
            "description": "Bytecode"
          }
        ],
        "responses": {
          "200": {
            "description": "TokenInfo",
            "schema": {
              "$ref": "#/definitions/TokenInfo"
            }
          }
        }
      }
    },
    "/file/append": {
      "post": {
        "tags": [
          "File controller"
        ],
        "summary": "Append a File.",
        "parameters": [
          {
            "in": "query",
            "required": true,
            "name": "text",
            "type": "string",
            "default": "Testing",
            "description": "Adding text to file..."
          },
          {
            "in": "query",
            "required": true,
            "name": "fileID",
            "type": "string",
            "default": "0.0.3327503",
            "description": "File ID..."
          }

        ],
        "responses": {
          "200": {
            "description": "File ID"
          }
        }
      }
    },
    "/file/getcontent": {
      "post": {
        "tags": [
          "File controller"
        ],
        "summary": "Get file content.",
        "parameters": [
          {
            "in": "query",
            "required": true,
            "name": "fileID",
            "type": "string",
            "default": "0.0.3327503",
            "description": "Get file content..."
          }
        ],
        "responses": {
          "200": {
            "description": "File ID"
          }
        }
      }
    },
    "/file/getinfo": {
      "post": {
        "tags": [
          "File controller"
        ],
        "summary": "Get file info.",
        "parameters": [
          {
            "in": "query",
            "required": true,
            "name": "fileID",
            "type": "string",
            "default": "0.0.3327503",
            "description": "Get file info..."
          }
        ],
        "responses": {
          "200": {
            "description": "File info"
          }
        }
      }
    }
  },

  "definitions": {
    "User": {
      "required": [
        "name",
        "_id",
        "companies"
      ],
      "properties": {
        "_id": {
          "type": "integer",
          "uniqueItems": true
        },
        "isPublic": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        },
        "books": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "amount": {
                "type": "number"
              }
            }
          }
        },
        "companies": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
    "TopicInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "object",
          "properties": {
            "shard": {
              "type": "number"
            },
            "realm": {
              "type": "number"
            },
            "topic": {
              "type": "number"
            }
          }
        },
        "topicMemo": {
          "type": "string"
        },
        "runningHash": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "sequenceNumber": {
          "type": "number"
        },
        "expirationTime": {
          "type": "string"
        },
        "adminKey": {
          "type": "string"
        },
        "submitKey": {
          "type": "string"
        },
        "autoRenewPeriod": {
          "type": "string"
        },
        "autoRenewAccountId": {
          "type": "string"
        }
      }
    },
    "TokenInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "object",
          "properties": {
            "tokenId": {
              "type": "string"
            },
            "tokenAddressSol": {
              "type": "string"
            },
            "totalSupplyLow": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}